+++ web/src/infra/gameInfo/GameModePickerCard.module.css
@@ -1,33 +1,3 @@
 .OccupancySelect {
   padding-bottom: 8px;
 }
-
-.CustomizationCard {
-  padding: 16px;
-  max-width: 90vw;
-  max-height: 90vh;
-  text-align: left;
- }
-
-.DialogClose {
-  float: right;
-  position: relative;
-  left: 12px;
-}
-
-.DialogTitleWrapper {
-  padding-bottom: 16px;
-}
-
-.DialogContentWrapper {
- text-align: left;
- overflow-y: auto;
- overflow-x: hidden;
- height: calc(100% - 48px);
-}
-
-.DialogTitle {
-  position: relative;
-  top: 9px;
-  padding-bottom: 8px;
-}
