+++ web/src/infra/settings/FullCustomization.module.css
@@ -1,0 +1,29 @@
+.CustomizationCard {
+  padding: 16px;
+  max-width: 90vw;
+  max-height: 90vh;
+  text-align: left;
+ }
+
+.DialogClose {
+  float: right;
+  position: relative;
+  left: 12px;
+}
+
+.DialogTitleWrapper {
+  padding-bottom: 16px;
+}
+
+.DialogContentWrapper {
+ text-align: left;
+ overflow-y: auto;
+ overflow-x: hidden;
+ height: calc(100% - 48px);
+}
+
+.DialogTitle {
+  position: relative;
+  top: 9px;
+  padding-bottom: 8px;
+}
